requirements.txt
Django==1.6.5
django-allauth==0.17.0
django-extensions==1.3.8
oauthlib==0.6.3
python-openid==2.2.5
requests==2.3.0
requests-oauthlib==0.4.1
six==1.7.2
